packages/functions/requirements.txt
sendgrid
google-cloud-pubsub
cloud-sql-python-connector[pg8000]
psycopg2-binary
python-dotenv
google-api-python-client
google-cloud-aiplatform
beautifulsoup4
requests
functions-framework
